+++ Content.Server/Implants/RadioImplantSystem.cs
@@ -1,7 +1,7 @@
-﻿using Content.Server.Radio.Components;
-using Content.Shared.Implants;
+﻿using Content.Shared.Implants;
 using Content.Shared.Implants.Components;
 using Robust.Shared.Containers;
+using Content.Shared.Radio.Components;
 
 namespace Content.Server.Implants;
