+++ Content.Server/Radio/EntitySystems/HeadsetSystem.cs
@@ -1,6 +1,5 @@
 using Content.Server.Chat.Systems;
 using Content.Server.Emp;
-using Content.Server.Radio.Components;
 using Content.Shared._Mono.Radio;
 using Content.Shared.Inventory.Events;
 using Content.Shared.Radio;
