+++ Content.Server/Silicons/Borgs/BorgSwitchableTypeSystem.cs
@@ -1,7 +1,7 @@
 using Content.Server.Inventory;
-using Content.Server.Radio.Components;
 using Content.Shared._CorvaxNext.Silicons.Borgs.Components;
 using Content.Shared.Inventory;
+using Content.Shared.Radio.Components;
 using Content.Shared.Silicons.Borgs;
 using Content.Shared.Silicons.Borgs.Components;
 using Robust.Shared.Prototypes;
