+++ Content.Shared/Radio/Components/ActiveRadioComponent.cs
@@ -1,8 +1,7 @@
-using Content.Shared.Radio;
 using Robust.Shared.GameStates;
 using Robust.Shared.Prototypes;
 
-namespace Content.Server.Radio.Components;
+namespace Content.Shared.Radio.Components;
 
 /// <summary>
 ///     This component is required to receive radio message events.
