+++ Content.Shared/Radio/Components/IntrinsicRadioReceiverComponent.cs
@@ -1,11 +1,11 @@
-namespace Content.Server.Radio.Components;
+using Robust.Shared.GameStates;
+
+namespace Content.Shared.Radio.Components;
 
 /// <summary>
 ///     This component allows an entity to directly translate radio messages into chat messages. Note that this does not
 ///     automatically add an <see cref="ActiveRadioComponent"/>, which is required to receive radio messages on specific
 ///     channels.
 /// </summary>
-[RegisterComponent]
-public sealed partial class IntrinsicRadioReceiverComponent : Component
-{
-}
+[RegisterComponent, NetworkedComponent]
+public sealed partial class IntrinsicRadioReceiverComponent : Component;
