+++ Content.Shared/Radio/Components/IntrinsicRadioTransmitterComponent.cs
@@ -1,16 +1,14 @@
-using Content.Server.Chat.Systems;
 using Content.Shared.Chat;
-using Content.Shared.Radio;
+using Robust.Shared.GameStates;
 using Robust.Shared.Prototypes;
-using Robust.Shared.Serialization.TypeSerializers.Implementations.Custom.Prototype.Set;
 
-namespace Content.Server.Radio.Components;
+namespace Content.Shared.Radio.Components;
 
 /// <summary>
 ///     This component allows an entity to directly translate spoken text into radio messages (effectively an intrinsic
 ///     radio headset).
 /// </summary>
-[RegisterComponent]
+[RegisterComponent, NetworkedComponent]
 public sealed partial class IntrinsicRadioTransmitterComponent : Component
 {
     [DataField]
