+++ Content.Shared/Radio/Components/RadioMicrophoneComponent.cs
@@ -1,44 +1,41 @@
-using Content.Server.Radio.EntitySystems;
+using Content.Shared.Radio.EntitySystems;
 using Content.Shared.Chat;
-using Content.Shared.Radio;
-using Robust.Shared.Serialization.TypeSerializers.Implementations.Custom.Prototype;
-using Robust.Shared.Serialization.TypeSerializers.Implementations.Custom.Prototype.List;
+using Robust.Shared.GameStates;
+using Robust.Shared.Prototypes;
 
-namespace Content.Server.Radio.Components;
+namespace Content.Shared.Radio.Components;
 
 /// <summary>
 ///     Listens for local chat messages and relays them to some radio frequency
 /// </summary>
-[RegisterComponent]
-[Access(typeof(RadioDeviceSystem))]
+[RegisterComponent, NetworkedComponent]
+[Access(typeof(SharedRadioDeviceSystem))]
 public sealed partial class RadioMicrophoneComponent : Component
 {
-    [ViewVariables(VVAccess.ReadWrite)]
-    [DataField("broadcastChannel", customTypeSerializer: typeof(PrototypeIdSerializer<RadioChannelPrototype>))]
-    public string BroadcastChannel = SharedChatSystem.CommonChannel;
+    [DataField]
+    public ProtoId<RadioChannelPrototype> BroadcastChannel = SharedChatSystem.CommonChannel;
 
-    [ViewVariables(VVAccess.ReadWrite)]
-    [DataField("listenRange")]
-    public int ListenRange  = 4;
+    [DataField]
+    public int ListenRange = 4;
 
-    [DataField("enabled")]
+    [DataField]
     public bool Enabled = false;
 
-    [DataField("powerRequired")]
+    [DataField]
     public bool PowerRequired = false;
 
     /// <summary>
     /// Whether or not interacting with this entity
     /// toggles it on or off.
     /// </summary>
-    [DataField("toggleOnInteract")]
+    [DataField]
     public bool ToggleOnInteract = true;
 
     /// <summary>
     /// Whether or not the speaker must have an
     /// unobstructed path to the radio to speak
     /// </summary>
-    [DataField("unobstructedRequired")]
+    [DataField]
     public bool UnobstructedRequired = false;
 
     // Nuclear-14
