+++ Content.Shared/Radio/Components/WearingHeadsetComponent.cs
@@ -1,13 +1,13 @@
-using Content.Server.Radio.EntitySystems;
+using Robust.Shared.GameStates;
 
-namespace Content.Server.Radio.Components;
+namespace Content.Shared.Radio.Components;
 
 /// <summary>
 ///     This component is used to tag players that are currently wearing an ACTIVE headset.
 /// </summary>
-[RegisterComponent]
+[RegisterComponent, NetworkedComponent]
 public sealed partial class WearingHeadsetComponent : Component
 {
-    [DataField("headset")]
+    [DataField]
     public EntityUid Headset;
 }
